refactor: change babel register to swc ts-node
