+++ packages/ordi/fetch.js
@@ -1,4 +1,4 @@
 const useFetch =
-  require("./build/shared/context/fetch/usecase/use-fetch").default;
+  require("./build/shared/context/fetch/usecase/use-fetch").useFetch;
 
 module.exports = { useFetch };
