+++ packages/ordi/package.json
@@ -83,13 +83,13 @@
     "@babel/preset-env": "^7.24.7",
     "@babel/preset-react": "^7.24.7",
     "@babel/preset-typescript": "^7.24.7",
-    "@babel/register": "^7.24.6",
     "@babel/runtime": "^7.24.7",
     "@loadable/babel-plugin": "^5.16.1",
     "@loadable/component": "^5.16.4",
     "@loadable/server": "^5.16.5",
     "@loadable/webpack-plugin": "^5.15.2",
     "@pmmmwh/react-refresh-webpack-plugin": "^0.5.15",
+    "@swc/core": "^1.7.0",
     "babel-loader": "^9.1.3",
     "chalk": "^5.3.0",
     "core-js": "^3.37.1",
@@ -107,6 +107,7 @@
     "react-router-dom": "5",
     "run-script-webpack-plugin": "^0.2.0",
     "style-loader": "^4.0.0",
+    "ts-node": "^10.9.2",
     "webpack": "^5.92.1",
     "webpack-cli": "^4.10.0",
     "webpack-dev-middleware": "^5.3.4",
