+++ packages/ordi/tsconfig.json
@@ -10,7 +10,16 @@
     "forceConsistentCasingInFileNames": true,
     "isolatedModules": true,
     "outDir": "./build",
-    "lib": ["es5", "es2015", "es2016", "es2017", "es2018", "es2019", "es2020", "dom"],
+    "lib": [
+      "es5",
+      "es2015",
+      "es2016",
+      "es2017",
+      "es2018",
+      "es2019",
+      "es2020",
+      "dom"
+    ],
     "maxNodeModuleJsDepth": 0,
     "module": "es2020",
     "moduleResolution": "node",
@@ -29,26 +38,21 @@
     "strict": true,
     "strictNullChecks": true,
     "target": "es6",
-    "plugins": [{ "transform": "typescript-transform-paths", "afterDeclarations": true }],
+    "plugins": [
+      { "transform": "typescript-transform-paths", "afterDeclarations": true }
+    ],
     "baseUrl": ".",
     "jsx": "react-jsx",
     "paths": {
-      "@BUILD_ROUTE" : ["src/client/routes/index.ts"],
+      "@BUILD_ROUTE": ["src/client/routes/index.ts"],
       "@BUILD_404": ["src/client/error/_404.tsx"],
-      "@BUILD_500": [
-        "src/client/error/_500.tsx"
-      ],
-      "@DOCUMENT" : ["src/client/_document.tsx"],
-      "@APP": [
-        "src/client/_app.tsx"
-      ],
+      "@BUILD_500": ["src/client/error/_500.tsx"],
+      "@DOCUMENT": ["src/client/_document.tsx"],
+      "@APP": ["src/client/_app.tsx"]
     },
     "typeRoots": ["./node_modules/@types"],
     "types": ["node"]
   },
-    "include": ["src/**/*"],
-    "exclude": [
-        "node_modules",
-        "build",
-    ]
+  "include": ["src/**/*", "utils/resolve.ts", "utils/ifDev.ts"],
+  "exclude": ["node_modules", "build"]
 }
